feat: added auth to pages
I have add some auth logic to the pages and also wrote the logic for logging in.
