+++ packages/api/src/router/user.ts
@@ -130,7 +130,8 @@ export const userRouter = {
         walletId: z.string(),
         name: z.string(),
         email: z.string().email(),
-        image: z.string(),
+        image: z.string().optional(),
+        bio: z.string().optional(),
         location: z.string().optional(),
         browser: z.string().optional(),
         operatingSystem: z.string().optional(),
