+++ turbo.json
@@ -63,7 +63,7 @@
     "NEXT_PUBLIC_DEPLOYMENT_ENVIRONMENT",
     "AUTH_SECRET",
     "PORT",
-
+    "NEXT_PUBLIC_THIRDWEB_AUTH_DOMAIN",
     "IPINFO_API_KEY",
     "THIRDWEB_ADMIN_PRIVATE_KEY",
     "THIRDWEB_AUTH_DOMAIN",
